Ensures model is marked as fitted
Sets the `fitted_` attribute to True after the model is trained. This ensures that subsequent calls to `predict_proba` will not attempt to use an untrained model.

Casts the `y_batch` to int8 to avoid potential overflow issues.

+++ mlex/models/lstm.py
@@ -118,5 +118,5 @@ class LSTM(nn.Module, BaseEstimator, ClassifierMixin):
         test_loader = self.create_test_loader(X, y)
         y_true = []
         for _, y_batch in test_loader:
-            y_true.extend(np.array(y_batch, dtype="int4").flatten())
+            y_true.extend(np.array(y_batch, dtype="int8").flatten())
         return y_true
